moving select2 out from viral
