+++ app/components/select2/v1/component.html.erb
