+++ test/components/previews/select2_component_preview.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Select2ComponentPreview < ViewComponent::Preview
+  def default; end
+end
