+++ test/components/previews/select2_component_preview/default.html.erb
