--- test/components/previews/viral_select2_component_preview.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class ViralSelect2ComponentPreview < ViewComponent::Preview
-  def default; end
-end
